moved logic to PresetManagerPresenter
