+++ NonMaps/src/com/nonlinearlabs/NonMaps/client/presenters/PresetManager.java
@@ -2,6 +2,11 @@ package com.nonlinearlabs.NonMaps.client.presenters;
 
 import java.util.ArrayList;
 
+import com.google.gwt.i18n.client.NumberFormat;
+import com.nonlinearlabs.NonMaps.client.NonMaps;
+import com.nonlinearlabs.NonMaps.client.dataModel.editBuffer.EditBufferModel;
+import com.nonlinearlabs.NonMaps.client.world.maps.presets.bank.preset.Preset;
+
 public class PresetManager {
 	public static class Bank {
 
@@ -21,5 +26,32 @@ public class PresetManager {
 		public String uuid;
 	}
 
+	public String getBaseLoadedPresetNumberString() {		
+		EditBufferModel eb = EditBufferModel.get();
+		String ret = "";
+		Preset preset = NonMaps.get().getNonLinearWorld().getPresetManager().findLoadedPreset();
+		com.nonlinearlabs.NonMaps.client.world.maps.presets.bank.Bank bank = preset != null ? preset.getParent() : null;
+		
+		if(eb.loadedPreset.getValue().equals("Init")) {
+			ret = "Init";
+		} else if(bank == null && preset == null) {
+			ret = "";
+		} else if(bank != null && preset != null) {
+			ret = bank.getOrderNumber() + "-" + NumberFormat.getFormat("000").format(preset.getNumber()); 
+		}
+
+		return ret;
+	}
+	
+	public String getLoadedPresetNumberString() {
+		boolean mod = NonMaps.theMaps.getNonLinearWorld().getParameterEditor().isModified();
+		String ret = getBaseLoadedPresetNumberString();
+		
+		if(ret.isEmpty())
+			return ret;
+		
+		return ret += mod ? " *" : "";
+	}
+	
 	public ArrayList<Bank> banks = new ArrayList<Bank>();
 }
